+++ src/CCalendar.h
@@ -27,7 +27,7 @@ public:
 
     [[nodiscard]] vector<shared_ptr<CEvent>> getSortedEvents(const shared_ptr<CEvent> &event) const;
 
-    [[nodiscard]] vector<shared_ptr<CEvent>> getDailyEvents(size_t year, size_t month, size_t day) const;
+    [[nodiscard]] vector<shared_ptr<CEvent>> getDailyEvents(int year, int month, int day) const;
 
     /**
      * Adds event to the calendar if possible. Implementation of the event->isConflict polymorphism
