+++ src/CDisplayCalendar.cpp
@@ -4,7 +4,7 @@ void CDisplayCalendar::setCalendar(const CCalendar &calendar) {
     mCalendar = calendar;
 }
 
-ostream &CDisplayCalendar::displayDailyEvents(ostream &out, size_t year, size_t month, size_t day) const {
+ostream &CDisplayCalendar::displayDailyEvents(ostream &out, int year, int month, int day) const {
     auto dailyEvents = mCalendar.getDailyEvents(year, month, day);
 
     printSeparator(out);
